eval/requirements-green.txt
# AgentX Green Agent Dependencies

# Async web framework (Flask-compatible)
quart>=0.19.0
quart-cors>=0.7.0
hypercorn>=0.16.0

# Legacy Flask support (for compatibility)
flask>=2.0.0
flask-cors>=4.0.0

# Core AgentX dependencies
sqlglot>=20.0.0
duckdb>=0.9.0

# PostgreSQL support (optional)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP client (async)
aiohttp>=3.9.0
requests>=2.28.0

# Resilience
tenacity>=8.2.0

# A2A Protocol Support
a2a-sdk[http-server]>=0.3.20
uvicorn>=0.38.0
